+++ cmd/sandbox/main.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"context"
+	"fmt"
 	"os"
 	"os/signal"
 
@@ -13,6 +14,7 @@ func main() {
 	defer stop()
 	go func() {
 		<-ctx.Done()
+		fmt.Fprintln(os.Stderr, "\nCancellation requested. Cleaning up; press Ctrl+C again to force exit.")
 		stop()
 	}()
 	os.Exit(cli.Run(ctx, os.Args[1:], os.Stdin, os.Stdout, os.Stderr))